Move the torchao filter test in with the other torchao tests
Review feedback from SunMarc. accelerate has no tests/quantization/torchao
directory, so the test goes to tests/test_fp8.py, which holds the existing
TestTorchAO cases, as a plain CPU class rather than a launcher one: the
layer swap itself needs no accelerator.
