+++ tests/test_utils.py
@@ -33,7 +33,6 @@ from accelerate.test_utils.testing import (
     require_non_cpu,
     require_non_torch_xla,
     require_torch_min_version,
-    require_torchao,
     require_tpu,
     require_triton,
     torch_device,
@@ -680,30 +679,3 @@ def test_purge_env_vars_restores_previous_values():
 
     del os.environ["ACCELERATE_SOME_ENV_VAR"]
     del os.environ["ACCELERATE_ANOTHER_ENV_VAR"]
-
-
-@require_torchao
-def test_convert_model_to_fp8_ao_skips_the_first_and_last_linear_layers():
-    # convert_model_to_fp8_ao documents that it converts every nn.Linear "except the first and
-    # last", and find_first_last_linear_layers exists because quantizing those two destabilises
-    # training. The default module_filter_func was filter_first_and_last_linear_layers, which
-    # re-derives the first and last linear from the module it is handed. torchao hands it one
-    # candidate layer at a time, so it always found ("", "") and filtered nothing, and every
-    # linear including the first and last was converted.
-    from torchao.float8.float8_linear import Float8Linear
-
-    from accelerate.utils.ao import convert_model_to_fp8_ao
-
-    class ToyModel(nn.Module):
-        def __init__(self):
-            super().__init__()
-            self.embed_proj = nn.Linear(32, 64)
-            self.block = nn.Sequential(nn.Linear(64, 64), nn.ReLU())
-            self.lm_head = nn.Linear(64, 32)
-
-    model = ToyModel()
-    convert_model_to_fp8_ao(model)
-
-    assert not isinstance(model.embed_proj, Float8Linear)
-    assert not isinstance(model.lm_head, Float8Linear)
-    assert isinstance(model.block[0], Float8Linear)
